Some refactoring to prepare for a new adapter
