+++ Gemfile
@@ -8,7 +8,7 @@ gemspec
 
 gem 'passenger',              '~> 5.2.1'
 gem 'mongoid',                '~> 6.4.0'
-gem 'devise',                 '~> 4.4.3'
+gem 'devise',                 '~> 4.6.2'
 gem 'sass-rails',             '~> 5.0'
 gem 'uglifier',               '>= 1.3.0'
 gem 'coffee-rails',           '~> 4.2'
