+++ lib/locomotivecms_search.rb
@@ -1,6 +1,7 @@
 require 'locomotive/search/errors'
 require 'locomotive/search/backend_factory'
 require 'locomotive/search/engine'
+require 'locomotive/search/backend/base'
 
 module Locomotive::Search
   # Your code goes here...
